fixing warnigns andd moved more stuff around

+++ Shake_Engine/src/Shake/Debug/Instrumentor.h
@@ -127,7 +127,7 @@ namespace Shake
             long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().
                 count();
 
-            const uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+            const uint32_t threadID = static_cast<uint32_t>(std::hash<std::thread::id>{}(std::this_thread::get_id()));
             Instrumentor::Get().WriteProfile({m_Name, start, end, threadID});
 
             m_Stopped = true;
